+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Runtime/Core/Controllers/CursorLocoMotionController.cs
@@ -1,7 +1,7 @@
+using System;
 using Studio23.SS2.BetterCursor.Data;
 using UnityEngine;
 using UnityEngine.InputSystem;
-using UnityEngine.InputSystem.LowLevel;
 
 namespace Studio23.SS2.BetterCursor.Core
 {
@@ -15,7 +15,6 @@ namespace Studio23.SS2.BetterCursor.Core
         private Vector2 _minScreenBounds;
         private Vector2 _maxScreenBounds;
 
-        private InputDevice _lastUsedDevice;
 
 
         private void Awake()
@@ -28,12 +27,7 @@ namespace Studio23.SS2.BetterCursor.Core
             _canvas = canvas;
             _cursorTransform.sizeDelta = cursorData.PixelSize;
             _cursorTransform.pivot = cursorData.HotSpot;
-        }
-
-        private void Start()
-        {
             SetupBounds();
-            SetupLastUsedDevice();
         }
 
         private void Update()
@@ -42,15 +36,15 @@ namespace Studio23.SS2.BetterCursor.Core
         }
 
 
-        public Vector2 GetCursorImagePosition()
+        internal Vector2 GetCursorImagePosition()
         {
-            return _cursorTransform.anchoredPosition;
+            return _cursorTransform.anchoredPosition* _canvas.scaleFactor;
         }
 
         private void UpdateCursorPosition()
         {
             var cursorPosition = CursorActionAsset["CursorPosition"].ReadValue<Vector2>();
-            if (IsController())
+            if (BetterCursor.Instance.IsController())
                 HandleControllerInput(cursorPosition);
             else
                 HandleMouseInput(cursorPosition);
@@ -73,37 +67,18 @@ namespace Studio23.SS2.BetterCursor.Core
                 _canvas.worldCamera,
                 out var localMousePosition
             );
-
             _cursorTransform.localPosition = localMousePosition;
         }
 
+
         private void SetupBounds()
         {
-            var canvas = _cursorTransform.GetComponentInParent<Canvas>();
-            float screenWidth = Screen.width;
-            float screenHeight = Screen.height;
-            var screenSizeInCanvas = new Vector2(screenWidth, screenHeight) / canvas.scaleFactor;
+            float screenWidth = Screen.width/ _canvas.scaleFactor;
+            float screenHeight = Screen.height/ _canvas.scaleFactor;
+            var screenSizeInCanvas = new Vector2(screenWidth, screenHeight);
             var uiElementSize = _cursorTransform.sizeDelta / 2.0f;
             _minScreenBounds = new Vector2(uiElementSize.x / 2, uiElementSize.y / 2);
             _maxScreenBounds = screenSizeInCanvas;
-        }
-
-        private void SetupLastUsedDevice()
-        {
-            InputSystem.onActionChange += (obj, change) =>
-            {
-                if (change == InputActionChange.ActionPerformed)
-                {
-                    var inputAction = (InputAction)obj;
-                    var lastControl = inputAction.activeControl;
-                    _lastUsedDevice = lastControl.device;
-                }
-            };
-        }
-
-        private bool IsController()
-        {
-            return !(_lastUsedDevice is Keyboard || _lastUsedDevice is Mouse);
         }
     }
 }